Api added + form edited

+++ api/user_resource.py
@@ -1,11 +1,12 @@
-from flask_restful import abort, Resource
-from flask import jsonify
 import data
+from flask_restful import Resource
+from flask import jsonify
 
 
 class UserResource(Resource):
     def get(self, user_id):
-        abort_if_users_not_found(user_id)
+        if abort_if_users_not_found(user_id):
+            return jsonify({'status': f"User {user_id} not found"})
         session = data.db_session.create_session()
         users = session.get(data.user.User, user_id)
         return jsonify({'user': {'address': users.address, 'birthday': users.birthday, 'email': users.email,
@@ -14,7 +15,8 @@ class UserResource(Resource):
                                  'profile_banner': users.profile_banner.hex()}})
 
     def delete(self, user_id):
-        abort_if_users_not_found(user_id)
+        if abort_if_users_not_found(user_id):
+            return jsonify({'status': f"User {user_id} not found"})
         session = data.db_session.create_session()
         users = session.get(data.user.User, user_id)
         session.delete(users)
@@ -26,4 +28,4 @@ def abort_if_users_not_found(user_id):
     session = data.db_session.create_session()
     users = session.get(data.user.User, user_id)
     if not users:
-        abort(404, message=f"User {user_id} not found")
+        return True
